When opening an item from search, link and activity buttons don't work Fixes #262
