+++ views/search.ejs
@@ -31,6 +31,7 @@
 
 <%- include('partials/commondialogs.ejs') %>
 <%- include('partials/itemview.ejs') %>
+<%- include('partials/itemactivity.ejs') %>
 
 <script defer type="module" src="/public/v<%= version %>/js/search.js"></script>
